create `coherence_goals` and `negation` modules
